pyproject.toml
[project]
name = "cortx-catalog-gen"
version = "0.1.0"
description = "Data Catalog & Semantic Layer Builder for Cortx"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "pandas>=2.0",
    "sqlalchemy>=2.0",
    "pyarrow>=12.0",
    "groq>=0.4",
    "sentence-transformers>=2.0",
    "python-dotenv>=1.0",
    "numpy>=1.24",
    "flask>=2.0",
    "flask-cors>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
kaggle = [
    "kaggle>=1.7",
]

[project.scripts]
cortx-catalog-gen = "cortx_catalog.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
